+++ utils/realloc_string.h
@@ -12,7 +12,7 @@
 #define REALLOC_STRING(format, value, pointer)                                 \
   do {                                                                         \
     int size = snprintf(NULL, 0, format, value);                               \
-    pointer = realloc(pointer, size + 1);                                      \
+    pointer = realloc(pointer, (size_t)size + 1);                              \
     sprintf(pointer, format, value);                                           \
   } while (0)
